Split up tests into specialized files
